refactor: create item-collection-repository directory
