+++ src/components/importModal.tsx
@@ -2,7 +2,7 @@ import React, { useState, useRef } from "react"
 import clsx from "clsx"
 import { ImportItemsAppendUsecase } from "../usecase/import-items-append-usecase"
 import { ImportItemsOverwriteUsecase } from "../usecase/import-items-overwrite-usecase"
-import { ItemCollectionRepository } from "../infra/itemCollectionRepository"
+import { ItemCollectionRepository } from "../infra/item-collection-repository/item-collection-repository"
 import { ChromeStorageClient } from "../infra/chrome-storage-client/chrome-storage-client"
 import { JSONFileClient } from "../infra/json-file-client/json-file-client"
 import { FileIOClient } from "../infra/file-io-client/file-io-client"
