+++ src/components/options/OptionsContainer.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useCallback } from "react"
 import { useItemCollection } from "../../domain/item/useItem"
 import { useItemInfo } from "../../domain/itemInfo/itemInfo"
 import { OptionsTemplate, Props } from "./OptionsTemplate"
-import { ItemCollectionRepository } from "../../infra/itemCollectionRepository"
+import { ItemCollectionRepository } from "../../infra/item-collection-repository/item-collection-repository"
 import { ChromeStorageClient } from "../../infra/chrome-storage-client/chrome-storage-client"
 import { useCustomBlock } from "../../domain/customBlock/useCustomBlock"
 import { useBlockHandlers } from "../../hooks/util/useBlockHandlers"
