+++ src/popup.tsx
@@ -4,7 +4,7 @@ import "bulma/css/bulma.css"
 import { PopupContainer } from "./components/popup/PopupContainer"
 import { useFetch } from "./shared/useFetch"
 import { ChromeStorageClient } from "./infra/chrome-storage-client/chrome-storage-client"
-import { ItemCollectionRepository } from "./infra/itemCollectionRepository"
+import { ItemCollectionRepository } from "./infra/item-collection-repository/item-collection-repository"
 import { GetCurrentItemUsecase } from "./usecase/get-current-item-usecase"
 
 function PopupPage(): JSX.Element {
